Tidied this up a bit to merge. More to do, but need to get this folded in.
